drop filter support
It was unreliable, filtered away only additional properties (did not
even support filtering additional items) and hid validation errors.

Also mutation of input objects is not a good practice, that is not
a way to create sanitized objects.

Also, filter had 0 test coverage - no tests depended on filtering.

Error message is added to keep errors in drop-in replacements visible.